refactor: moved layer tab components (except maplibre ones) to pages/map/layers folder
