+++ sites/geohub/src/components/controls/vector-styles/TextMaxWidth.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { LayerSpecification } from 'maplibre-gl';
 	import { createEventDispatcher, getContext } from 'svelte';
-	import NumberInput from '$components/controls/NumberInput.svelte';
+	import NumberInput from '$components/util/NumberInput.svelte';
 	import type { Layer } from '$lib/types';
 	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
