+++ sites/geohub/src/components/pages/map/layers/LegendColorMapRow.svelte
