+++ sites/geohub/src/components/pages/map/layers/LegendTypeSwitcher.svelte
