+++ sites/geohub/src/components/pages/map/layers/header/DeleteMenu.svelte
