+++ sites/geohub/src/components/pages/map/layers/header/LayerHeader.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-	import DataCardInfoMenu from '$components/controls/DataCardInfoMenu.svelte';
-	import DeleteMenu from '$components/controls/DeleteMenu.svelte';
-	import VisibilityButton from '$components/controls/VisibilityButton.svelte';
-	import Legend from '$components/controls/vector-styles/Legend.svelte';
+	import DataCardInfoMenu from '$components/pages/map/layers/header/DataCardInfoMenu.svelte';
+	import DeleteMenu from '$components/pages/map/layers/header/DeleteMenu.svelte';
+	import VisibilityButton from '$components/pages/map/layers/header/VisibilityButton.svelte';
+	import Legend from '$components/pages/map/layers/header/Legend.svelte';
 	import { AccessLevel } from '$lib/config/AppConfig';
 	import { clean, getAccessLevelIcon, getLayerStyle, handleEnterKey, initTippy } from '$lib/helper';
 	import type { Layer, RasterTileMetadata, VectorTileMetadata } from '$lib/types';
