+++ sites/geohub/src/components/pages/map/layers/header/Legend.svelte
