+++ sites/geohub/src/components/pages/map/layers/raster/RasterDefaultLegend.svelte
