+++ sites/geohub/src/components/pages/map/layers/raster/RasterHistogram.svelte
