+++ sites/geohub/src/components/pages/map/layers/raster/RasterOpCat.svelte
