+++ sites/geohub/src/components/pages/map/layers/raster/RasterTransformAdvanced.svelte
@@ -14,7 +14,7 @@ A component designed to apply where expression to a raster layer through titiler
 */
 	import Step from '$components/util/Step.svelte';
 	import Wizard from '$components/util/Wizard.svelte';
-	import RasterTransformNumbersInput from '$components/controls/RasterTransformNumbersInput.svelte';
+	import RasterTransformNumbersInput from '$components/pages/map/layers/raster/RasterTransformNumbersInput.svelte';
 	import { RasterArithmeticOperators, RasterComparisonOperators } from '$lib/config/AppConfig';
 	import { fetchUrl, getActiveBandIndex, getLayerSourceUrl, handleEnterKey } from '$lib/helper';
 	import type { BandMetadata, Layer, RasterLayerStats, RasterTileMetadata } from '$lib/types';
