+++ sites/geohub/src/components/pages/map/layers/vector/OperationButtons.svelte
