+++ sites/geohub/src/components/pages/map/layers/vector/ValueInput.svelte
