+++ sites/geohub/src/components/pages/map/layers/vector/VectorFilterOperators.svelte
