+++ sites/geohub/src/components/pages/map/layers/vector/VectorHeatmap.svelte
