+++ sites/geohub/src/components/pages/map/layers/vector/VectorPolygon.svelte
