+++ sites/geohub/src/components/pages/map/layers/vector/VectorSymbol.svelte
