+++ sites/geohub/src/components/util/NumberInput.svelte
