+++ sites/geohub/src/components/util/NumberInput.test.ts
@@ -1,7 +1,7 @@
 import { describe, beforeEach, expect, it, vi } from 'vitest';
 import { cleanup, fireEvent, render, type RenderResult } from '@testing-library/svelte';
 
-import NumberInput from '$components/controls/NumberInput.svelte';
+import NumberInput from '$components/util/NumberInput.svelte';
 
 beforeEach(cleanup);
