+++ sites/geohub/src/routes/(map)/map/+layout.svelte
@@ -2,7 +2,7 @@
 	import { browser } from '$app/environment';
 	import { beforeNavigate } from '$app/navigation';
 	import { page } from '$app/stores';
-	import Content from '$components/Content.svelte';
+	import Content from '$components/pages/map/Content.svelte';
 	import Header from '$components/header/Header.svelte';
 	import Notification from '$components/util/Notification.svelte';
 	import { fromLocalStorage, isStyleChanged, storageKeys, toLocalStorage } from '$lib/helper';
