--- sites/geohub/src/styles/button-icons-selected.scss
@@ -1,9 +1,0 @@
-.icon-selected {
-  opacity: 0.5;
-  display: inline;
-  cursor: pointer;
-
-  &:hover {
-    opacity: 1;
-  }
-}
